requirements.txt
Flask==3.1.2
pandas==2.3.2
gunicorn==23.0.0
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0